Add a POC compartmentalization library.
This adds a proof-of-concept hybrid compartmentalization library for
CheriBSD (for the `morello-hybrid` target).

The aim of the library is to provide tooling for compartmentalizing existing C
programs with respect to code and data.

The library aims to support:
* execution compartmentalization: the PCC of each compartment is restricted
  according to the bounds established (statically) using the `__comp(COMP_NAME)`
  macro
* data compartmentalization: compartments have separate and non-overlapping
  data regions (TODO: this restriction will need to be relaxed to support data
  sharing between compartments)

The implementation is heavily inspired by [`inter_comp_call` hybrid
compartmentalization examples] from the CapableVMs [cheri-examples]
repository. It doesn't add anything new other than the use of sections
to delineate compartment boundaries (for execution
compartmentalization), and support for compartments with multiple entry
points.

See README.md for more details.

[`inter_comp_call` hybrid compartmentalization examples]: https://github.com/capablevms/cheri-examples/blob/d6ecb72aaa479f6126a69b9e70de1ec2fc49cdc0/hybrid/compartment_examples/inter_comp_call/malicious_compartments/shared/switch_compartment.S
[cheri-examples]: https://github.com/capablevms/cheri-examples/tree/d6ecb72aaa479f6126a69b9e70de1ec2fc49cdc0